Change decrement logic
Decrements the quantity of second by the difference of the current time and the last tab date.

The decrements by the heartbeat was causing a bug in our application. The session of our users is 90 minutes. Some of them was putting their computer in sleep and then try to continue their session when they return because the pop-up was not showing up.
This implementation prevents that.